Module lutario::basis

Basis manipulation.

We make heavy use of block-diagonal matrices here. Conceptually, a block-diagonal matrix is equivalent to a list of blocks, which need not be the same size nor do they have to be square. Many operations on block-diagonal matrices translate to just operations on each block, which can be easily parallelized in principle.

"Channels" are labels that uniquely identifies each block.

"Auxiliary" are labels that identify each row or column within a block.

The combination of channel and auxiliary must, by definition, uniquely identify a state.

We further subdivide each block into subblocks. This is needed for, say, separating the particle vs hole states.

"Parts" are labels that identify the subblocks along one of the axes.

The use of parts has both pros and cons. The advantage is that things like diagonalization of a block matrix is very simple because each block can be diagonalized independently. The disadvantage is that it adds another layer of complication and overhead.

Naming convention:

l = channel (in text, λ is sometimes used)

u = auxiliary

x = part ("eXcitation" / unoccupancy)

p = state index, isomorphic to (l, u)

i = state ID, isomorphic to state index but in channelized order

These letters could refer to either indices or the concrete objects. When there is ambiguity, indicial variables are prefixed with i.

Be aware that these labels may also have other meanings in more concrete contexts (e.g. l can mean orbital angular momentum, and p can mean parity).

Modules

occ
Convenient aliases for occupancies.

Structs

BasisChart
A basis chart contains information needed to interpret an axis of a block-diagonal matrix. The type is parametrized by:

BasisLayout
A basis layout has only enough information to understand the structure of a matrix along one axis, but not to interpret its elements.

CACHE
Global cache for storing basis information.

ChanState
Channelized state index.

Fence
Iterator over Option<impl Increment>.

HashChart
A chart is a bijection between a zero-based interval (i.e. indices) and an arbitrary set of objects of type T. HashChart is an implementation of this idea using HashMap.

HashChartInsertResult

Hashed

MatChart

MatLayout

Orb
Orbital ID i, in channelized order. Not to be confused with orbital index.

OrbIx
Orbital index p, in the original basis order. Not to be confused with orbital ID.

PackedOptChanState

PartState

Enums

Occ
Occupancy of an orbital (single-particle state).

Occ20
Occupancy of antisymmetrized two-particle states.

Traits

FromUsize

Increment
Associates data types with sequences.

IntoUsize

Functions

siphash128

Type Definitions

State2