Struct lutario::basis::BasisChart

pub struct BasisChart<L, X, U> {
    pub chan_chart: HashChart<L, u32>,
    pub part_chart: HashChart<X, u32>,
    pub aux_encoder: FnvHashMap<(u32, U), u32>,
    pub aux_decoder: Box<[U]>,
    pub orb_from_ix: Box<[Orb]>,
    pub orb_to_ix: Box<[OrbIx]>,
    pub layout: BasisLayout,
}

A basis chart contains information needed to interpret an axis of a block-diagonal matrix. The type is parametrized by:

L (λ): channel type,

X (χ): part type, and

U (μ): auxiliary type.

Fields
chan_chart: HashChart<L, u32> part_chart: HashChart<X, u32> aux_encoder: FnvHashMap<(u32, U), u32>

(l, μ) → u

aux_decoder: Box<[U]>

(p ≡ chan_offset(l) + u) → μ

orb_from_ix: Box<[Orb]>

p → i

orb_to_ix: Box<[OrbIx]>

i → p

layout: BasisLayout
Methods

impl<L, X, U> BasisChart<L, X, U> where
L: Hash + Eq + Clone,
X: Hash + Eq + Clone,
U: Hash + Eq + Clone,

pub fn new<I>(states: I) -> Self where
I: Iterator<Item = PartState<X, ChanState<L, U>>>,

pub fn new_with<I>(
states: I,
chan_chart: HashChart<L, u32>,
part_chart: HashChart<X, u32>
) -> Self where
I: Iterator<Item = PartState<X, ChanState<L, U>>>,

Create a BasisChart with the given states, using the provided charts as a starting point. The charts will be augmented if necessary.

impl<L, X, U> BasisChart<L, X, U>

pub fn layout(&self) -> &BasisLayout

pub fn decode_chan(&self, l: u32) -> Option<&L>

pub fn decode_part(&self, x: u32) -> Option<&X>

pub fn decode_aux(&self, l: u32, u: u32) -> Option<&U>

impl<L: Hash + Eq, X, U> BasisChart<L, X, U>

pub fn encode_chan(&self, l: &L) -> Option<u32>

impl<L, X: Hash + Eq, U> BasisChart<L, X, U>

pub fn encode_part(&self, x: &X) -> Option<u32>

impl<L, X, U: Hash + Eq> BasisChart<L, X, U>

pub fn encode_aux(&self, l: u32, u: &U) -> Option<u32>