Struct lutario::basis::BasisLayout

pub struct BasisLayout {
    pub num_parts: u32,
    pub part_offsets: Box<[u32]>,
    pub chan_offsets: Box<[u32]>,
    pub state_chans: Box<[u32]>,
}

A basis layout has only enough information to understand the structure of a matrix along one axis, but not to interpret its elements.

Fields
num_parts: u32 part_offsets: Box<[u32]>

part_offsets[l * num_parts + x - 1] == ∑[x' < x] num_states_l_x(l, x') where l < nl && x > 0 && x < num_parts

chan_offsets: Box<[u32]>

chan_offsets[l] == ∑[l' < l] num_states_l(l') where l ≤ nl

state_chans: Box<[u32]>

states[p] == l where p ≤ np

Methods

impl BasisLayout

pub fn num_states(&self) -> u32

pub fn num_chans(&self) -> u32

pub fn num_parts(&self) -> u32

pub fn num_auxs(&self, l: u32) -> u32

pub fn chan_dim(&self, l: u32) -> u32

pub fn part_dim(&self, l: u32, x: u32) -> u32

pub fn channelize(&self, p: u32) -> ChanState

pub fn dechannelize(&self, lu: ChanState) -> u32

pub fn chan_offset(&self, l: u32) -> u32

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn chan_offsets(&self) -> &[u32]

pub fn part_offset(&self, l: u32, x: u32) -> u32

pub fn part_offsets(&self) -> MatRef<u32>