Struct lutario::basis::HashChart

pub struct HashChart<T, I = usize> {
    pub encoder: FnvHashMap<T, I>,
    pub decoder: Vec<T>,
}

A chart is a bijection between a zero-based interval (i.e. indices) and an arbitrary set of objects of type T. HashChart is an implementation of this idea using HashMap.

Fields
encoder: FnvHashMap<T, I> decoder: Vec<T>
Methods

impl<T: Hash + Eq, I> HashChart<T, I>

pub fn encode<Q>(&self, q: &Q) -> Option<&I> where
Q: Hash + Eq,
T: Borrow<Q>,

impl<T, I: FromUsize> HashChart<T, I>

pub fn len(&self) -> I

impl<T, I: IntoUsize> HashChart<T, I>

pub fn decode(&self, i: I) -> Option<&T>

impl<T: Hash + Eq, I: FromUsize> HashChart<T, I>

pub fn reorder_by_key<K, F, G>(&mut self, key: F, permut: G) where
K: Ord,
F: FnMut(&T) -> K,
G: FnMut(&I, &I),

Reorder the elements in the HashChart via some total ordering, changing the association between elements and indices. The callback receives the permutation that maps from old indices to new.

impl<T: Hash + Eq + Clone, I: FromUsize + Clone> HashChart<T, I>

pub fn insert(&mut self, t: T) -> HashChartInsertResult<I>