Struct lutario::block::Bd

pub struct Bd<M>(pub Vec<M>);

A block-diagonal matrix is a vector of matrix blocks.

This wrapper is necessary to avoid conflicting implementations of the vector traits.

Methods

impl<T> Bd<Mat<T>>

pub fn extent_mat_as_tri(&self) -> usize

impl<T: Clone> Bd<Mat<T>>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn clone_mat_to_tri_slice<'a>(&self, a: &'a mut [T]) -> &'a mut [T]

Pack the lower triangles of the matrix blocks into the given array and then return the remaining part of the array.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn clone_mat_from_tri_slice<'a, S: Trs<T>>(
&mut self,
trs: &S,
a: &'a [T]
) -> &'a [T]