Struct lutario::block_mat::RawBlockMatRef

pub struct RawBlockMatRef { /* fields omitted */ }

A block-diagonal matrix with unspecified lifetime and unspecified type.

Methods

impl RawBlockMatRef

pub unsafe fn new(
ptr: *mut (),
num_blocks: usize,
block_offsets: *const usize,
block_strides: *const usize,
block_num_rows: *const usize,
block_num_cols: *const usize
) -> Self

Unsafe preconditions:

ptr must point to an array with sufficient length.

All the indexing arrays must be at least as long as num_blocks.

pub unsafe fn as_ptr<T>(self) -> *const T

Unsafe preconditions: T must be the correct type.

pub unsafe fn mut_ptr<T>(self) -> *mut T

Unsafe preconditions: T must be the correct type.

pub fn num_blocks(self) -> usize

Unsafety: No attempt is made to constrain the lifetime.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn block_offsets<'a>(self) -> &'a [usize]

Unsafety: No attempt is made to constrain the lifetime.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn block_strides<'a>(self) -> &'a [usize]

Unsafety: No attempt is made to constrain the lifetime.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn block_num_rows<'a>(self) -> &'a [usize]

Unsafety: No attempt is made to constrain the lifetime.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn block_num_cols<'a>(self) -> &'a [usize]

Unsafety: No attempt is made to constrain the lifetime.

pub unsafe fn shape_at(self, l: usize) -> ValidMatShape

pub unsafe fn offset_at(self, l: usize) -> usize