Struct lutario::half::Half

pub struct Half<T>(pub T);

Type that logically behaves like half-integers, but what is actually stored is twice its logical value.

For example, Half(3) represents the fraction 3/2.

Methods

impl<T> Half<T>

pub fn twice(self) -> T

Unwrap to twice its logical value.

impl<T: Add<Output = T> + Clone> Half<T>

pub fn double(self) -> Self

impl<T: Signed> Half<T>

pub fn abs(self) -> Half<T>

impl<T: Signed + Ord> Half<T>

pub fn in_multiplet_of(self, j: Self) -> bool

impl<T: Clone + Div<Output = T> + Rem<Output = T> + Zero + One> Half<T>

pub fn try_get(self) -> Result<T, Self>

Get the value if it's half-even. Otherwise, returns Err(self).

impl<T> Half<T> where
T: Clone + Debug + Div<Output = T> + Rem<Output = T> + Zero + One,

pub fn unwrap(self) -> T

Equivalent to try_get().unwrap().

impl<T: Ord + Sub> Half<T>

pub fn abs_diff(self, other: Self) -> Half<T::Output>

impl<T: Add<Output = T> + Sub<Output = T> + One + Ord + Clone> Half<T>

Important traits for RangeInclusive<A>
impl<A> Iterator for RangeInclusive<A> where
A: Clone + PartialOrd + Zero + One + ToPrimitive, type Item = A;

pub fn tri_range(self, other: Self) -> RangeInclusive<Self>

Obtain the range of values that satisfy the triangular condition, i.e. the range from |self − other| to self + other (inclusive).

Important traits for RangeInclusive<A>
impl<A> Iterator for RangeInclusive<A> where
A: Clone + PartialOrd + Zero + One + ToPrimitive, type Item = A;

pub fn tri_range_2(
(j1, j2): (Self, Self),
(j3, j4): (Self, Self)
) -> RangeInclusive<Self>

impl<T: Neg<Output = T> + Clone> Half<T>

Important traits for RangeInclusive<A>
impl<A> Iterator for RangeInclusive<A> where
A: Clone + PartialOrd + Zero + One + ToPrimitive, type Item = A;

pub fn multiplet(self) -> RangeInclusive<Self>

impl Half<i32>

pub fn phase(self) -> f64

Returns the phase (-1)^j.

Panics if self is half-odd.

pub fn weight(self, exponent: i32) -> f64

Returns (2 j + 1)^(exponent / 2).

Important traits for RangeInclusive<A>
impl<A> Iterator for RangeInclusive<A> where
A: Clone + PartialOrd + Zero + One + ToPrimitive, type Item = A;

pub fn quad_range(self, b: Self, c: Self) -> RangeInclusive<Self>

Obtain the range of values that satisfy the quadrangular condition.