Module lutario::imsrg

Implementation of IM-SRG(2)

This module contains all the terms of IM-SRG(2) commutator or linked product, implemented for J-scheme. It also contains a set of helper routines for driving the IM-SRG evolution (Run).

Structs

Conf
Configuration for an IM-SRG(2) run

Run
The state of an IM-SRG(2) run

Enums

DenomType

Error
Error type for Run.

Functions

c011
Term 011

c022
Term 022

c111
Term 111

c112
Term 112

c121
Term 121

c212
Term 212

c221
Term 221

c1220
Term 1220

c1221
Term 1221

c2220
Term 2220

c2221
Term 2221 (particle-hole term)

c2222
Term 2222

c2222_base
Generic 2220/2222-type term

commut
Commutator of two many-body operators

linked
Linked product of two many-body operators

monopole_elem
Compute the monopole matrix element.

white_gen
Calculate the White generator, multiply by alpha, and then add it to eta.