Struct lutario::io::parser::Parser

pub struct Parser<R> { /* fields omitted */ }

Adds infinite lookahead to a Read stream and the ability to save and restore parsing state arbitrarily.

Methods

impl<R> Parser<R>

pub fn new(reader: R) -> Self

pub fn with_capacity(reader: R, buf_cap: usize) -> Self

The buf_cap sets the capacity of a single buffer chunk. This doesn't affect the size of the lookahead (it's always infinite), as multiple chunks are permitted; this only affects the efficiency.

pub fn save(&self) -> State

Save the parser state for later. While the state object lives, it will store all the bytes between the saved point and the current parser point in memory, so it's important to dispose of the state when it's no longer needed.

pub fn restore(&mut self, state: State)

Restore the parser state.

impl<R: Read> Parser<R>

pub fn refill(&mut self) -> Result<()>

pub fn get(&mut self) -> Result<Option<u8>>

pub fn match_pred<F>(&mut self, pred: F) -> Result<Option<u8>> where
F: FnOnce(u8) -> bool,

pub fn match_bytes(&mut self, s: &[u8]) -> Result<bool>

pub fn munch_whitespace<'a>(&mut self) -> Result<bool>

pub fn munch_space<'a>(&mut self) -> Result<bool>

pub fn get_token<'a>(&mut self, s: &'a mut String) -> Result<&'a mut String>

Get the next non-whitespace token. This may return an empty string.

pub fn get_line<'a>(&mut self, line: &'a mut String) -> Result<&'a mut String>

Get the next line, including the line terminator. Returns the empty string at EOF.

pub fn next_line<'a>(
&mut self,
line: &'a mut String,
index: &mut usize
) -> Result<&'a mut String>