Module lutario::linalg

Linear algebra things

Most of things here are just thin wrappers over BLAS and LAPACK functionality along with some traits to enable polymorphism.

Modules

lamch
Defines and re-exports some floating-point constants following the LAPACK convention. Basically, this module contains anything you would otherwise obtain using lamch.

Structs

AdjSym

Enums

EigenvalueRange
Desired range of eigenvalues.

Part

Transpose

Constants

ANTIHERMITIAN

ANTISYMMETRIC

HERMITIAN

SYMMETRIC

Traits

Conj

Gemm

Heevr

NormSqr

Functions

gemm
A thin wrapper over Gemm::gemm that panics if the buffers are too small or if the sizes don't match. We omit Layout because it can be trivially emulated by exchanging a and b.

heevr
If left is true, calculate left eigenvectors stored as rows; otherwise, calculate right eigenvectors stored as columns.

heevr_n
If left is true, calculate left eigenvalues; otherwise, calculate right eigenvalues

mat_axpby
y ← α × x + β × y

part_to_u8