Module lutario::linalg::lamch::f32

Constants

BASE
The radix or base of the internal representation of f32.

EMAX
Maximum possible power of 2 exponent.

EMIN
One greater than the minimum possible normal power of 2 exponent.

EPS
Relative machine epsilon according to the LAPACK convention. Equal to half of std::*::EPSILON.

PREC
Difference between 1.0 and the next largest representable number.

RMAX
Largest finite f32 value.

RMIN
Smallest positive normal f32 value.

RND
Whether proper rounding (true) or chopping (false) occurs in addition.

SFMIN
Safe minimum such that 1.0 / SFMIN does not overflow.

T
Number of significant digits in base 2.