Struct lutario::mat::MatMut

pub struct MatMut<'a, T: 'a> { /* fields omitted */ }

Methods

impl<'a, T> MatMut<'a, T>

pub fn new(slice: &mut &'a mut [T], shape: ValidMatShape) -> Option<Self>

Split *slice into two parts: the first part becomes the matrix, while the second part is stored back in slice. If the slice is too short, None is returned.

pub unsafe fn from_raw(ptr: *mut T, shape: ValidMatShape) -> Self

pub fn shape(&self) -> ValidMatShape

pub fn num_rows(&self) -> usize

pub fn num_cols(&self) -> usize

pub fn dims(&self) -> (usize, usize)

pub fn stride(&self) -> usize

pub fn into_ref(self) -> MatRef<'a, T>

pub fn as_ref(&self) -> MatRef<T>

pub fn as_mut(&mut self) -> MatMut<T>

pub fn as_ptr(&self) -> *mut T

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn index(self, i: usize, j: usize) -> &'a mut T

pub fn get(self, i: usize, j: usize) -> Option<&'a mut T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn get_unchecked(self, i: usize, j: usize) -> &'a mut T

pub unsafe fn offset_unchecked(&mut self, i: usize, j: usize) -> *mut T

Important traits for MatRowsMut<'a, T>
impl<'a, T> Iterator for MatRowsMut<'a, T> type Item = &'a mut [T];

pub fn rows(self) -> MatRowsMut<'a, T>

pub fn row(self, i: usize) -> Option<&'a mut [T]>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn row_unchecked(self, i: usize) -> &'a mut [T]

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn to_slice(self) -> &'a mut [T]

Unsafe because the slice includes padding elements as well.

pub fn slice(self, is: Range<usize>, js: Range<usize>) -> Self

pub fn split_at_row(self, i: usize) -> (Self, Self)

pub fn split_at_col(self, j: usize) -> (Self, Self)

impl<'m, T: Clone> MatMut<'m, T>

pub fn fill(&mut self, value: &T)

pub fn clone_from_ref(&mut self, source: MatRef<T>)

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn clone_from_tri_slice<'a, S: Trs<T>>(
&mut self,
trs: &S,
a: &'a [T]
) -> &'a [T]

impl<'a, T: Clone> MatMut<'a, T>

pub fn clone_from_trs_mat<S>(&mut self, source: &TrsMat<S, T>) where
S: Trs<T>,