Struct lutario::mat::MatRef

pub struct MatRef<'a, T: 'a> { /* fields omitted */ }

Methods

impl<'a, T> MatRef<'a, T>

pub fn new(slice: &mut &'a [T], shape: ValidMatShape) -> Option<Self>

Split *slice into two parts: the first part becomes the matrix, while the second part is stored back in slice. If the slice is too short, None is returned.

pub unsafe fn from_raw(ptr: *const T, shape: ValidMatShape) -> Self

pub fn shape(self) -> ValidMatShape

pub fn num_rows(self) -> usize

pub fn num_cols(self) -> usize

pub fn dims(self) -> (usize, usize)

pub fn stride(self) -> usize

pub fn extent(self) -> usize

pub fn as_ptr(self) -> *const T

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn index(self, i: usize, j: usize) -> &'a T

pub fn get(self, i: usize, j: usize) -> Option<&'a T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn get_unchecked(self, i: usize, j: usize) -> &'a T

pub unsafe fn offset_unchecked(self, i: usize, j: usize) -> *const T

Important traits for MatRows<'a, T>
impl<'a, T> Iterator for MatRows<'a, T> type Item = &'a [T];

pub fn rows(self) -> MatRows<'a, T>

pub fn row(self, i: usize) -> Option<&'a [T]>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn row_unchecked(self, i: usize) -> &'a [T]

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn to_slice(self) -> &'a [T]

Unsafe because the slice includes padding elements as well.

pub fn slice(self, is: Range<usize>, js: Range<usize>) -> Self

pub fn split_at_row(self, i: usize) -> (Self, Self)

pub fn split_at_col(self, j: usize) -> (Self, Self)

pub fn extent_as_tri(&self) -> usize

impl<'m, T: Clone> MatRef<'m, T>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn clone_to_tri_slice<'a>(self, a: &'a mut [T]) -> &'a mut [T]

Clone the lower triangle of the matrix into the given array and then return the remaining part of the array.