Module lutario::nuclei

Nuclei systems.

Here we use the particle physics convention of proton = +½, neutron = −½.

Modules

darmstadt
Darmstadt ME2J matrix element format.

vrenorm
VRenormalize matrix element format used in CENS MBPT.

Structs

Ho3dIter
Iterator for 3D harmonic oscillator states in (e, l, j)-order.

Ho3dModTrunc
Modified 3D harmonic oscillator basis truncation

Ho3dTrunc
Truncation scheme for a 3D harmonic oscillator basis

JNpjwKey

JtwNpjKey

Nj
Principal quantum number and total angular momentum magnitude

Nlj
Principal quantum number, orbital angular momentum magnitude, and total angular momentum magnitude

Npj
Principal quantum number, parity, and total angular momentum magnitude

Npjmw
Principal quantum number, parity, total angular momentum magnitude, total angular momentum projection, and isospin projection.

Npjw
Principal quantum number, parity, total angular momentum magnitude, and isospin projection.

Nucleons

Nucleus

ORB_ANG_CHART

OrbAng
Orbital angular momentum magnitude (l)

Pj
Parity and total angular momentum magnitude

Pjtw
Parity, total angular momentum magnitude, isospin magnitude, and isospin projection.

Pmw
Parity, total angular momentum projection, and isospin projection

Pw
Parity and isospin projection

SimpleNucleus

Functions

kinetic_ho3d_mat_elem
Calculate the kinetic energy matrix element in a 3D harmonic oscillator basis:

make_ho3d_op_j

make_ho3d_op_m

make_ke_op_j

make_ke_op_m

make_v_op_j
Load two-body matrix elements from the given hash table. Note that this function only uses canonicalized keys.

op1_j_to_m

op2_j_to_m