Module lutario::plane_wave_basis

3D plane-wave basis.

The “harmonic number” n of a 3D plane wave is defined as:

n = k L / (2 π)

where L is the box length and k is the angular wave number.

Structs

HarmSpin

HarmSpinIso

HarmTable

Functions

is_sum_of_3_squares
Test if this integer can be expressed as a sum of three squares using Legendre’s three-square theorem.