Module lutario::tri_mat

Row-major, lower-triangular matrices (i ≥ j).

Modules

trs

Structs

StrictTriMatDim
Dimensions of a strict lower triangular matrix, equal to the number of rows or columns.

TriMat

TriMatDim
Dimensions of a non-strict lower triangular matrix, equal to the number of rows or columns.

TriMatMut

TriMatRef

TriMatRows

TriMatRowsMut

TrsMat
Matrices that possess symmetry under transposition, represented as triangular matrices.

Traits

Trs
Transpose symmetry.