Struct lutario::tri_mat::TriMatMut

pub struct TriMatMut<'a, T: 'a> { /* fields omitted */ }

Methods

impl<'a, T> TriMatMut<'a, T>

pub fn new(slice: &mut &'a mut [T], shape: TriMatDim) -> Option<Self>

Split *slice into two parts: the first part becomes the matrix, while the second part is stored back in slice. If the slice is too short, None is returned.

pub unsafe fn from_raw(ptr: *mut T, shape: TriMatDim) -> Self

pub fn shape(&self) -> TriMatDim

pub fn into_ref(self) -> TriMatRef<'a, T>

pub fn as_ref(&self) -> TriMatRef<T>

pub fn as_mut(&mut self) -> TriMatMut<T>

pub fn as_ptr(&self) -> *mut T

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn index(self, i: usize, j: usize) -> &'a mut T

pub fn get(self, i: usize, j: usize) -> Option<&'a mut T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn get_unchecked(self, i: usize, j: usize) -> &'a mut T

pub unsafe fn offset_unchecked(&mut self, i: usize, j: usize) -> *mut T

Important traits for TriMatRowsMut<'a, T>
impl<'a, T> Iterator for TriMatRowsMut<'a, T> type Item = &'a mut [T];

pub fn rows(self) -> TriMatRowsMut<'a, T>

pub fn row(self, i: usize) -> Option<&'a mut [T]>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn row_unchecked(self, i: usize) -> &'a mut [T]

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn to_slice(self) -> &'a mut [T]

impl<'a, T: Clone> TriMatMut<'a, T>

pub fn fill(&mut self, value: &T)

pub fn clone_from_ref(&mut self, source: TriMatRef<T>)