Struct lutario::tri_mat::TriMatRef

pub struct TriMatRef<'a, T: 'a> { /* fields omitted */ }

Methods

impl<'a, T> TriMatRef<'a, T>

pub fn new(slice: &mut &'a [T], shape: TriMatDim) -> Option<Self>

Split *slice into two parts: the first part becomes the matrix, while the second part is stored back in slice. If the slice is too short, None is returned.

pub unsafe fn from_raw(ptr: *const T, shape: TriMatDim) -> Self

pub fn shape(self) -> TriMatDim

pub fn as_ptr(self) -> *const T

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn index(self, i: usize, j: usize) -> &'a T

pub fn get(self, i: usize, j: usize) -> Option<&'a T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn get_unchecked(self, i: usize, j: usize) -> &'a T

pub unsafe fn offset_unchecked(self, i: usize, j: usize) -> *const T

Important traits for TriMatRows<'a, T>
impl<'a, T> Iterator for TriMatRows<'a, T> type Item = &'a [T];

pub fn rows(self) -> TriMatRows<'a, T>

pub fn row(self, i: usize) -> Option<&'a [T]>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn row_unchecked(self, i: usize) -> &'a [T]

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn to_slice(self) -> &'a [T]