Module lutario::vector_driver

Generalizes the idea of vector operations to enable transparent parallelization.

A vector driver is a device that understands how to perform map-reduce-like operations on its own family of vector-like objects. It is the interface used by the sg-ode solver and is quite powerful, albeit awkward to use. The vec_apply! macro helps a bit.

The documentation here is very incomplete. There is a bit more information in the docs of the sg-ode vector driver interface in C, after which the Rust interface is modelled.

The MPI vector driver will be implemented some day, but for now you can only use the basic vector driver. There is nothing difficult about the MPI implementation – proof of concepts have already been made. The main concern is that it adds a really heavy dependency (MPI) to Lutario, so it probably should be done in a separate crate.

Modules

basic
The most trivial vector driver.

c
Vector driver interface with C.

Traits

VectorDriver

Functions

assert_all_eq